[MOD] Short jwt test back as UT
