+++ back/src/test/java/com/openclassrooms/starterjwt/security/jwt/JwtUtilsTest.java
@@ -1,6 +1,7 @@
 package com.openclassrooms.starterjwt.security.jwt;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
@@ -11,6 +12,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.boot.test.system.CapturedOutput;
+import org.springframework.boot.test.system.OutputCaptureExtension;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.test.util.ReflectionTestUtils;
@@ -62,6 +65,68 @@ public class JwtUtilsTest {
         boolean valid = jwtUtils.validateJwtToken(token);
         // THEN
         assertTrue(valid);
+    }
+
+    @Test
+    @ExtendWith(OutputCaptureExtension.class)
+    public void givenExpiredToken_validateJwtToken_ShouldReturnFalseAndLogIt(CapturedOutput capture) {
+        // GIVEN
+        String token = SigningHelper.sign("bob@test.com", -1, "jwtKey");
+        // WHEN
+        boolean valid = jwtUtils.validateJwtToken(token);
+        // THEN
+        assertFalse(valid);
+        assertThat(capture.getOut()).contains("JWT token is expired: ");
+    }
+
+    @Test
+    @ExtendWith(OutputCaptureExtension.class)
+    public void givenInvalidToken_validateJwtToken_ShouldReturnFalseAndLogIt(CapturedOutput capture) {
+        // GIVEN
+        String token = "fyJhbGciOiJIUzUxMiJ9.eyJz9ZSmR01GH_3nMUM2hqloKKLOJAFF0LoI6a-cITi88jjtF3fOBjVDrZWjVyqzYG9jL-7YVOsSFrAKWAvnw";
+        // WHEN THEN
+        // WHEN
+        boolean valid = jwtUtils.validateJwtToken(token);
+        // THEN
+        assertFalse(valid);
+        assertThat(capture.getOut()).contains("Invalid JWT token: ");
+    }
+
+    @Test
+    @ExtendWith(OutputCaptureExtension.class)
+    public void givenBadSignedToken_validateJwtToken_ShouldReturnFalseAndLogIt(CapturedOutput capture) {
+        // GIVEN
+        String token = SigningHelper.sign("bob@test.com", 5000, "badJwtKey");
+        // WHEN
+        boolean valid = jwtUtils.validateJwtToken(token);
+        // THEN
+        assertFalse(valid);
+        assertThat(capture.getOut()).contains("Invalid JWT signature: ");
+    }
+
+    @Test
+    @ExtendWith(OutputCaptureExtension.class)
+    public void givenUnsupportedToken_validateJwtToken_ShouldReturnFalseAndLogIt(CapturedOutput capture) {
+        // GIVEN
+        String token = SigningHelper.unsigned("bob@test.com", 5000, "jwtKey");
+        // WHEN
+        boolean valid = jwtUtils.validateJwtToken(token);
+        // THEN
+        assertFalse(valid);
+        assertThat(capture.getOut()).contains("JWT token is unsupported: ");
+    }
+
+    @Test
+    @ExtendWith(OutputCaptureExtension.class)
+    public void givenEmptyClaimToken_validateJwtToken_ShouldReturnFalseAndLogIt(CapturedOutput capture)
+            throws Exception {
+        // GIVEN
+        String token = "";
+        // WHEN
+        boolean valid = jwtUtils.validateJwtToken(token);
+        // THEN
+        assertFalse(valid);
+        assertThat(capture.getOut()).contains("JWT claims string is empty: ");
     }
 
 }
